Adding support to delete an ApiToken if it is no longer being used
