+++ src/Adapters/TokenResourceAdapter.php
@@ -9,6 +9,7 @@ use Rhubarb\Crown\Response\NotAuthorisedResponse;
 use Rhubarb\RestApi\Adapters\ModelResourceAdapter;
 use Rhubarb\Scaffolds\TokenBasedRestApi\Model\ApiToken;
 use Rhubarb\Scaffolds\TokenBasedRestApi\Resources\TokenResource;
+use Rhubarb\Stem\Filters\Equals;
 
 class TokenResourceAdapter extends ModelResourceAdapter
 {
@@ -49,5 +50,45 @@ class TokenResourceAdapter extends ModelResourceAdapter
         unset($resource->id);
 
         return $resource;
+    }
+
+
+    public function delete($payload, $params, ?WebRequest $request)
+    {
+        $token = $this->extractAuthorizationToken($request);
+
+        try {
+            $apiToken = ApiToken::findFirst(new Equals('Token', $token));
+            $apiToken->delete();
+        } catch (RecordNotFoundException $ex) {
+        }
+
+        $response = new \stdClass();
+        $response->status = true;
+
+        return $response;
+    }
+
+    private function extractAuthorizationToken(WebRequest $request)
+    {
+        $token = '';
+        if (!$request->header("Authorization")) {
+            Log::debug("Authorization header missing. If using fcgi be sure to instruct Apache to include this header", "RESTAPI");
+            return $token;
+        }
+
+        $authString = trim($request->header("Authorization"));
+
+        if (stripos($authString, "token") !== 0 && stripos($authString, "bearer") !== 0) {
+            return $token;
+        }
+
+        if (!preg_match("/(token|bearer)(=|\s+)\"?([[:alnum:]]+)\"?/i", $authString, $match)) {
+            return $token;
+        }
+
+        $token = $match[3];
+
+        return $token;
     }
 }
